Don't use proxy_app_verify_callback()
The callback seems to only be used to set the check_issued callback, which can
be set directly from the code. As per https://github.com/openssl/openssl/pull/1294
it's not possible to be set in STORE_CTX but only in STORE, so let's hope it
will get used when actually neeeded.